Improve handling of current bingo data
`GetCurrentBingo` and `SetCurrentBingo` DB requests now work with `i64`
instead of `u32` for timestamps (which is the same datatype as `chrono`
uses) and `GetCurrentBingo` now returns a full `Bingo` struct instead of
just the ID as `u8`. Additionally, calling `update_current_bingo` to
fetch updated data from the API now returns `start` and `end` timestamps
instead of a `bool` indicating whether the bingo had ended.

+++ src/shared/db/write.rs
@@ -45,8 +45,8 @@ impl DbRequest for AddBingoMapping {
 
 pub struct SetCurrentBingo {
     pub bingo_id: u8,
-    pub start: u32,
-    pub end: u32,
+    pub start: i64,
+    pub end: i64,
 }
 impl DbRequest for SetCurrentBingo {
     type ReturnValue = Result<()>;
